# platform/api/openapi.yaml
openapi: 3.1.0
info:
  title: Terraform Agent Platform API
  version: 0.1.0
  description: |
    TAP control-plane v1 contract. Conventions: cursor pagination
    (`cursor`/`limit` query params, `next_cursor` in envelopes), RFC 7807
    `application/problem+json` errors, `Idempotency-Key` header on POSTs,
    UUIDs serialized as strings, UTC ISO 8601 timestamps, OIDC bearer authn.
servers:
  - url: /v1
security:
  - oidc: []
...
paths:
...
  /agents:
    post:
      tags: [agents]
      summary: Register an agent (or a new version)
      operationId: registerAgent
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, version, domain, manifest]
              properties:
                name: { type: string }
                version: { type: string, description: semver }
                domain: { $ref: "#/components/schemas/AgentDomain" }
                manifest: { type: object }
                artifact_uri: { type: [string, "null"] }
      responses:
        "201":
          description: Registered agent.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        default: { $ref: "#/components/responses/Problem" }
    get:
      tags: [agents]
      summary: Discover agents
      operationId: listAgents
      parameters:
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
        - { name: domain, in: query, schema: { $ref: "#/components/schemas/AgentDomain" } }
        - { name: capability, in: query, schema: { type: string } }
      responses:
        "200": { description: Cursor page of agents. }
        default: { $ref: "#/components/responses/Problem" }
  /agents/{agent_id}:
    get:
      tags: [agents]
      summary: Get an agent
      operationId: getAgent
      parameters:
        - { name: agent_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The agent.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Agent" }
        default: { $ref: "#/components/responses/Problem" }
  /agents/{agent_id}/versions:
    get:
      tags: [agents]
      summary: List agent versions
      operationId: listAgentVersions
      parameters:
        - { name: agent_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: Cursor page of agent versions. }
        default: { $ref: "#/components/responses/Problem" }
  /marketplace/listings:
    get:
      tags: [marketplace]
      summary: Browse/search marketplace listings
      operationId: listListings
      parameters:
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
        - { name: tier, in: query, schema: { type: string, enum: [internal, community, enterprise] } }
        - { name: domain, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200": { description: Cursor page of listings. }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [marketplace]
      summary: Publish a signed OCI agent artifact
      operationId: publishListing
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [artifact_uri]
              properties:
                artifact_uri: { type: string }
                tier: { type: string, enum: [internal, community, enterprise] }
                description: { type: [string, "null"] }
      responses:
        "201": { description: Created listing (cosign-verified). }
        default: { $ref: "#/components/responses/Problem" }
  /marketplace/listings/{listing_id}:
    get:
      tags: [marketplace]
      summary: Get a listing
      operationId: getListing
      parameters:
        - { name: listing_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: The listing. }
        default: { $ref: "#/components/responses/Problem" }
  /marketplace/listings/{listing_id}/install:
    post:
      tags: [marketplace]
      summary: Install a listed agent into the tenant registry
      operationId: installListing
      parameters:
        - { name: listing_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      responses:
        "202": { description: Install accepted. }
        default: { $ref: "#/components/responses/Problem" }
  /policies:
    post:
      tags: [policies]
      summary: Create a policy set
      operationId: createPolicySet
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "201": { description: Created policy set. }
        default: { $ref: "#/components/responses/Problem" }
    get:
      tags: [policies]
      summary: List policy sets
      operationId: listPolicySets
      parameters:
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200": { description: Cursor page of policy sets. }
        default: { $ref: "#/components/responses/Problem" }
  /policies/{policy_set_id}:
    get:
      tags: [policies]
      summary: Get a policy set
      operationId: getPolicySet
      parameters:
        - { name: policy_set_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: The policy set. }
        default: { $ref: "#/components/responses/Problem" }
  /policies/evaluate:
    post:
      tags: [policies]
      summary: Evaluate a plan JSON ad hoc against applicable policy sets
      operationId: evaluatePolicies
      responses:
        "200": { description: "Aggregated OPA decision: allowed, enforcement_level, violations." }
        default: { $ref: "#/components/responses/Problem" }
  /approvals:
    get:
      tags: [approvals]
      summary: List approvals
      operationId: listApprovals
      parameters:
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
        - { name: approval_status, in: query, schema: { type: string, enum: [pending, approved, rejected, expired] } }
      responses:
        "200": { description: Cursor page of approvals. }
        default: { $ref: "#/components/responses/Problem" }
  /approvals/{approval_id}:
    get:
      tags: [approvals]
      summary: Get an approval
      operationId: getApproval
      parameters:
        - { name: approval_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: The approval. }
        default: { $ref: "#/components/responses/Problem" }
  /approvals/{approval_id}/decision:
    post:
      tags: [approvals]
      summary: Approve or reject; signals the waiting RunWorkflow
      operationId: decideApproval
      parameters:
        - { name: approval_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approve]
              properties:
                approve: { type: boolean }
                comment: { type: [string, "null"], maxLength: 2000 }
      responses:
        "200": { description: Decided approval. }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://login.example.com/tap/.well-known/openid-configuration
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque cursor from a previous page.
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string }
      description: Repeated POSTs with the same key return the original result.
...
  responses:
    Problem:
      description: RFC 7807 error.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [title, status]
      properties:
        type: { type: string, default: "about:blank" }
        title: { type: string }
        status: { type: integer }
        detail: { type: [string, "null"] }
        instance: { type: [string, "null"] }
...
    AgentDomain:
      type: string
      enum: [appops, devops, secops, netops, dataops, llmops, identity, costops]
    Agent:
      type: object
      required: [id, tenant_id, name, domain, enabled, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        domain: { $ref: "#/components/schemas/AgentDomain" }
        latest_version: { type: [string, "null"], description: semver }
        capabilities:
          type: array
          items: { type: string }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }